solving-pacman-backend/requirements.txt
black==23.9.1 ; python_version >= "3.10" and python_version < "3.13"
blinker==1.6.2 ; python_version >= "3.10" and python_version < "3.13"
cfgv==3.4.0 ; python_version >= "3.10" and python_version < "3.13"
click==8.1.7 ; python_version >= "3.10" and python_version < "3.13"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.13" and (sys_platform == "win32" or platform_system == "Windows")
coverage==7.3.2 ; python_version >= "3.10" and python_version < "3.13"
distlib==0.3.7 ; python_version >= "3.10" and python_version < "3.13"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.12.4 ; python_version >= "3.10" and python_version < "3.13"
flake8==6.1.0 ; python_version >= "3.10" and python_version < "3.13"
flask==3.0.0 ; python_version >= "3.10" and python_version < "3.13"
gunicorn==21.2.0 ; python_version >= "3.10" and python_version < "3.13"
identify==2.5.30 ; python_version >= "3.10" and python_version < "3.13"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "3.13"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "3.13"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "3.13"
markupsafe==2.1.3 ; python_version >= "3.10" and python_version < "3.13"
mccabe==0.7.0 ; python_version >= "3.10" and python_version < "3.13"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.13"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "3.13"
numpy==1.26.1 ; python_version >= "3.10" and python_version < "3.13"
packaging==23.2 ; python_version >= "3.10" and python_version < "3.13"
pathspec==0.11.2 ; python_version >= "3.10" and python_version < "3.13"
pillow==10.1.0 ; python_version >= "3.10" and python_version < "3.13"
platformdirs==3.11.0 ; python_version >= "3.10" and python_version < "3.13"
pluggy==1.3.0 ; python_version >= "3.10" and python_version < "3.13"
pre-commit==3.4.0 ; python_version >= "3.10" and python_version < "3.13"
pycodestyle==2.11.0 ; python_version >= "3.10" and python_version < "3.13"
pydocstyle-gitlab-code-quality==0.0.1 ; python_version >= "3.10" and python_version < "3.13"
pydocstyle==6.3.0 ; python_version >= "3.10" and python_version < "3.13"
pyflakes==3.1.0 ; python_version >= "3.10" and python_version < "3.13"
pytest-pydocstyle==2.3.2 ; python_version >= "3.10" and python_version < "3.13"
pytest==7.4.2 ; python_version >= "3.10" and python_version < "3.13"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.13"
setuptools==68.2.2 ; python_version >= "3.10" and python_version < "3.13"
snowballstemmer==2.2.0 ; python_version >= "3.10" and python_version < "3.13"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.8.0 ; python_version >= "3.10" and python_version < "3.11"
virtualenv==20.24.5 ; python_version >= "3.10" and python_version < "3.13"
werkzeug==3.0.0 ; python_version >= "3.10" and python_version < "3.13"
